move analytics code to AnalyticsClient.kt single session id for all FPTI events
